requirements.txt
python-dotenv
openai
openai_function_call

